[internal/color] ToNRGBA() delivers uint8 instead of int
That saves a lot of error prone integer conversion at the caller sides
and will save even more with additional refactoring at the callers.

+++ dialog/base.go
@@ -215,7 +215,7 @@ func (renderer *themedBackgroundRenderer) Objects() []fyne.CanvasObject {
 
 func (renderer *themedBackgroundRenderer) Refresh() {
 	r, g, b, _ := col.ToNRGBA(theme.Color(theme.ColorNameOverlayBackground))
-	bg := &color.NRGBA{R: uint8(r), G: uint8(g), B: uint8(b), A: 230}
+	bg := &color.NRGBA{R: r, G: g, B: b, A: 230}
 	renderer.rect.FillColor = bg
 }
